Restructure the C binding pipeline
