+++ .gitignore
@@ -5,6 +5,8 @@ __pycache__/
 
 # C extensions
 *.so
+petlib/_petlib.o
+petlib/_petlib.c
 
 # Distribution / packaging
 .Python
