+++ .travis.yml
@@ -2,19 +2,22 @@ language: python
 
 python:
   - "2.7"
-  - "3.5" 
+  - "3.6" 
+
+env:
+  - OPENSSL_VERSION=1.0
+  - OPENSSL_VERSION=1.1
 
 before_install:
   - sudo apt-get update -qq
   - sudo apt-get install -qq python-dev libssl-dev libffi-dev
-  - sudo apt-get install --reinstall libssl1.0.0
   - wget https://bootstrap.pypa.io/ez_setup.py -O - | sudo python
 
 # command to install dependencies
 install: 
+  - if [ "$OPENSSL_VERSION" = "1.1" ]; then sh petlib/_cffi_src/install_openssl_v1_1.sh > /dev/null 2>&1; fi
   - pip install tox-travis
   - pip install -r requirements.txt
-  - python setup.py install
 
 # command to run tests
 script: tox
