+++ petlib/cipher.py
@@ -3,15 +3,14 @@ from .bindings import _FFI, _C
 import pytest
 
 
-_pool = []
+def _check(return_val):
+    """Checks the return code of the C calls"""
+    if isinstance(return_val, int) and return_val == 1:
+      return
+    if isinstance(return_val, bool) and return_val == True:
+      return
 
-def get_intptr():
-    if _pool == []:
-        _pool.append( _FFI.new("int *") )
-    return _pool.pop()
-
-def return_intptr(ptr):
-    _pool.append(ptr)
+    raise Exception("Cipher exception: Unknown type %s or value %s" % (str(type(return_val)), str(return_val)))
 
 class Cipher(object):
     """ A class representing a symmetric cipher and mode.
@@ -38,13 +37,11 @@ class Cipher(object):
         True
 
     """
-
-    __slots__ = ["alg", "gcm", "_pool"]
+        
+    __slots__ = ["alg", "gcm"]
 
     def __init__(self, name, _alg=None):
         """Initialize the cipher by name."""
-
-        self._pool = []
 
         if _alg:
             self.alg = _alg
@@ -59,25 +56,27 @@ class Cipher(object):
 
         if "gcm" in name.lower():
             self.gcm = True
-
+        
         if "ccm" in name.lower():
             raise Exception("CCM mode not supported")
 
     def len_IV(self):
         """Return the Initialization Vector length in bytes."""
-        return int(self.alg.iv_len)
+        return int(_C.EVP_CIPHER_iv_length(self.alg))
     def len_key(self):
         """Return the secret key length in bytes."""
-        return int(self.alg.key_len)
+        return int(_C.EVP_CIPHER_key_length(self.alg))
     def len_block(self):
         """Return the block size in bytes."""
-        return int(self.alg.block_size)
+        return int(_C.EVP_CIPHER_block_size(self.alg))
     def get_nid(self):
         """Return the OpenSSL nid of the cipher and mode."""
-        return int(self.alg.nid)
+        return int(_C.EVP_CIPHER_nid(self.alg))
 
-    def op(self, key, iv=None, enc=1):
-        """Initializes a cipher operation, either encrypt or decrypt
+
+
+    def op(self, key, iv, enc=1):
+        """Initializes a cipher operation, either encrypt or decrypt 
         and returns a CipherOperation object
 
         Args:
@@ -86,47 +85,37 @@ class Cipher(object):
             enc (int): set to 1 to perform encryption, or 0 to perform decryption.
 
         """
-
-        if iv is None:
-            iv = _FFI.NULL
-
-        ok = True
-
-        if self._pool == []:
-            c_op = CipherOperation(enc)
-        else:
-            c_op = self._pool.pop()
-            c_op.init(enc)
-
-        ok &= ( len(key) == int(self.alg.key_len) )
-        ok &= ( enc == 0 or enc == 1 )
-
-        if not ok: raise Exception("Cipher exception: Wrong key length or enc mode.")
-
+        c_op = CipherOperation(enc)
+        _check( len(key) == self.len_key())
+        _check( enc in [0,1] )
+       
         if not self.gcm:
-            if iv != _FFI.NULL:
-                ok &= ( len(iv) == self.len_IV() )
-            ok &= ( _C.EVP_CipherInit_ex(c_op.ctx, self.alg, _FFI.NULL, key, iv, enc) )
+            _check( len(iv) == self.len_IV())
+            _check( _C.EVP_CipherInit_ex(c_op.ctx, 
+                self.alg,  _FFI.NULL, key, iv, enc) )
 
         else:
+            
 
-            ok &= ( _C.EVP_CipherInit_ex(c_op.ctx, self.alg,  _FFI.NULL, _FFI.NULL, _FFI.NULL, enc) )
+            _check( _C.EVP_CipherInit_ex(c_op.ctx, 
+                self.alg,  _FFI.NULL, _FFI.NULL, _FFI.NULL, enc) )
 
             # assert len(iv) <= self.len_block()
 
-            ok &= ( _C.EVP_CIPHER_CTX_ctrl(c_op.ctx, _C.EVP_CTRL_GCM_SET_IVLEN, len(iv), _FFI.NULL))
+            _check( _C.EVP_CIPHER_CTX_ctrl(c_op.ctx, 
+                _C.EVP_CTRL_GCM_SET_IVLEN, len(iv), _FFI.NULL))
 
             _C.EVP_CIPHER_CTX_ctrl(c_op.ctx, _C.EVP_CTRL_GCM_SET_IV_FIXED, -1, iv);
             _C.EVP_CIPHER_CTX_ctrl(c_op.ctx, _C.EVP_CTRL_GCM_IV_GEN, 0, iv)
 
-            ok &= ( _C.EVP_CipherInit_ex(c_op.ctx, _FFI.NULL,  _FFI.NULL, key, iv, enc) )
+            _check( _C.EVP_CipherInit_ex(c_op.ctx, 
+                _FFI.NULL,  _FFI.NULL, key, iv, enc) )
 
-        if not ok: raise Exception("Cipher exception: Init failed.")
         c_op.cipher = self
         return c_op
 
     def enc(self, key, iv):
-        """Initializes an encryption engine with the cipher with a specific key and Initialization Vector (IV).
+        """Initializes an encryption engine with the cipher with a specific key and Initialization Vector (IV). 
         Returns the CipherOperation engine.
 
         Args:
@@ -137,7 +126,7 @@ class Cipher(object):
         return self.op(key, iv, enc=1)
 
     def dec(self, key, iv):
-        """Initializes a decryption engine with the cipher with a specific key and Initialization Vector (IV).
+        """Initializes a decryption engine with the cipher with a specific key and Initialization Vector (IV). 
         Returns the CipherOperation engine.
 
         Args:
@@ -147,8 +136,8 @@ class Cipher(object):
         """
         return self.op(key, iv, enc=0)
 
-    #def __del__(self):
-    #    pass
+    def __del__(self):
+        pass
 
 # --------- AES GCM special functions ---------------
 
@@ -178,7 +167,7 @@ class Cipher(object):
             assoc (str): associated data that will be integrity protected, but not encrypted.
             tagl (int): the length of the tag, up to the block length.
 
-        Example:
+        Example: 
             Use of `quick_gcm_enc` and `quick_gcm_dec` for AES-GCM operations.
 
             >>> from os import urandom      # Secure OS random source
@@ -202,7 +191,7 @@ class Cipher(object):
         return (ciphertext, tag)
 
     def quick_gcm_dec(self, key, iv, cip, tag, assoc=None):
-        """One operation GCM decrypt. See usage example in "quick_gcm_enc".
+        """One operation GCM decrypt. See usage example in "quick_gcm_enc". 
         Throws an exception on failure of decryption
 
         Args:
@@ -216,16 +205,16 @@ class Cipher(object):
         dec = self.dec(key, iv)
         if assoc:
             dec.update_associated(assoc)
-
+        
         dec.set_tag(tag)
         plain = dec.update(cip)
-
+        
         try:
             plain += dec.finalize()
         except:
             raise Exception("Cipher: decryption failed.")
         return plain
-
+                
 
 class CipherOperation(object):
 
@@ -233,67 +222,30 @@ class CipherOperation(object):
 
     def __init__(self, xenc):
         self.ctx = _C.EVP_CIPHER_CTX_new()
-        self.init(xenc)
-
-    def init(self, xenc):
-        _C.EVP_CIPHER_CTX_init(self.ctx)
+        # _C.EVP_CIPHER_CTX_init(self.ctx)
         self.cipher = None
         self.xenc = xenc
-
-
-    def set_padding(self, pad):
-        """Sets the padding on or off, accodring to pad (bool).
-
-        Example:
-            >>> from os import urandom
-            >>> aes = Cipher("AES-128-ECB")     # Init AES in Electronic codebook mode
-            >>> key = urandom(16)
-            >>> iv  = None
-            >>>
-            >>> # Get a CipherOperation object for encryption
-            >>> enc = aes.enc(key, iv)
-            >>> enc.set_padding(False)
-            >>> ref = b"A" * 16
-            >>> ciphertext = enc.update(ref)
-            >>> ciphertext += enc.finalize()
-            >>> len(ciphertext)
-            16
-            >>> # Get a CipherOperation object for decryption
-            >>> dec = aes.dec(key, iv)
-            >>> dec.set_padding(False)
-            >>> plaintext = dec.update(ciphertext)
-            >>> plaintext += dec.finalize()
-            >>> plaintext == ref # Check resulting plaintest matches referece one.
-            True
-
-        """
-        ok = _C.EVP_CIPHER_CTX_set_padding(self.ctx, pad)
-        if not ok: raise Exception("Cipher exception: Set padding failed.")
-
+            
     def update(self, data):
         """Processes some data, and returns a partial result."""
-        block_len = self.cipher.alg.block_size # self.cipher.len_block()
+        block_len = self.cipher.len_block()
         alloc_len = len(data) + block_len + 1
-        # outl = _FFI.new("int *")
-        outl = get_intptr()
+        outl = _FFI.new("int *")
         outl[0] = alloc_len
         out = _FFI.new("unsigned char[]", alloc_len)
-
-        ok = _C.EVP_CipherUpdate(self.ctx, out, outl, data, len(data))
-        if not ok: raise Exception("Cipher exception: Update failed.")
-
+        
+        _check( _C.EVP_CipherUpdate(self.ctx, out, outl, data, len(data)) )
+        
         ret = bytes(_FFI.buffer(out)[:int(outl[0])])
-
-        return_intptr(outl)
         return ret
 
     def finalize(self):
-        """Finalizes the operation and may return some additional data.
+        """Finalizes the operation and may return some additional data. 
         Throws an exception if the authenticator tag is different from the expected value.
-
+        
         Example:
             Example of the exception thrown when an invalid tag is provided.
-
+            
             >>> from os import urandom
             >>> aes = Cipher.aes_128_gcm()              # Define an AES-GCM cipher
             >>> iv = urandom(16)
@@ -308,9 +260,9 @@ class CipherOperation(object):
             ... except:
             ...    print("Failure")
             Failure
-
+            
             Throws an exception since integrity check fails due to the invalid tag.
-
+        
         """
         block_len = self.cipher.len_block()
         alloc_len = block_len
@@ -319,9 +271,7 @@ class CipherOperation(object):
         out = _FFI.new("unsigned char[]", alloc_len)
 
         try:
-            ok = _C.EVP_CipherFinal_ex(self.ctx, out, outl)
-            if not ok: raise Exception("Cipher exception: Finalize failed.")
-
+            _check( _C.EVP_CipherFinal_ex(self.ctx, out, outl) )
             if outl[0] == 0:
                 return b''
 
@@ -337,14 +287,13 @@ class CipherOperation(object):
             self.set_tag(b"\00" * 16)
 
         outl = _FFI.new("int *")
-        ok  = ( _C.EVP_CipherUpdate(self.ctx, _FFI.NULL, outl, data, len(data)))
-        ok &=( outl[0] == len(data) )
-        if not ok: raise Exception("Cipher exception: Update associated data failed.")
+        _check( _C.EVP_CipherUpdate(self.ctx, _FFI.NULL, outl, data, len(data)))
+        _check( outl[0] == len(data) )
 
     def get_tag(self, tag_len = 16):
         """Get the GCM authentication tag. Execute after finalizing the encryption.
 
-        Example:
+        Example: 
             AES-GCM encryption usage:
 
             >>> from os import urandom
@@ -359,11 +308,11 @@ class CipherOperation(object):
 
         """
         tag = _FFI.new("unsigned char []", tag_len)
-        ok =  _C.EVP_CIPHER_CTX_ctrl(self.ctx, _C.EVP_CTRL_GCM_GET_TAG, tag_len, tag)
-        if not ok: raise Exception("Cipher exception: Cipher control failed.")
-
-        ret = bytes(_FFI.buffer(tag)[:])
-        return ret
+        ret =  _C.EVP_CIPHER_CTX_ctrl(self.ctx, _C.EVP_CTRL_GCM_GET_TAG, tag_len, tag)
+        _check( ret )
+        s = bytes(_FFI.buffer(tag)[:])
+        return s
+        
 
     def set_tag(self, tag):
         """Specify the GCM authenticator tag. Must be done before finalizing decryption
@@ -376,20 +325,15 @@ class CipherOperation(object):
             >>> dec = aes.dec(key=b"A"*16, iv=b"A"*16)  # Get a decryption CipherOperation
             >>> dec.update_associated(b"Hello")         # Feed in the non-secret assciated data.
             >>> plaintext = dec.update(ciphertext)      # Feed in the ciphertext for decryption.
-            >>> dec.set_tag(tag)                        # Provide the AES-GCM tag for integrity.
+            >>> dec.set_tag(tag)                        # Provide the AES-GCM tag for integrity. 
             >>> nothing = dec.finalize()                # Check and finalize.
             >>> assert plaintext == b'World!'
 
         """
-        ok = (_C.EVP_CIPHER_CTX_ctrl(self.ctx, _C.EVP_CTRL_GCM_SET_TAG, len(tag), tag))
-        if not ok: raise Exception("Cipher exception: Set tag failed.")
+        _check( _C.EVP_CIPHER_CTX_ctrl(self.ctx, _C.EVP_CTRL_GCM_SET_TAG, len(tag), tag))
 
     def __del__(self):
-        if self not in self.cipher._pool:
-            self.cipher._pool.append(self)
-        else:
-            _C.EVP_CIPHER_CTX_cleanup(self.ctx)
-            _C.EVP_CIPHER_CTX_free(self.ctx)
+        _C.EVP_CIPHER_CTX_free(self.ctx)
 
 
 ## When testing ignore extra variables
@@ -465,7 +409,6 @@ def test_aes_gcm_encrypt():
 
     tag = enc.get_tag(16)
     assert len(tag) == 16
-    assert isinstance(tag, bytes)
 
 def test_aes_gcm_encrypt_192():
     aes = Cipher.aes_192_gcm()
@@ -623,22 +566,5 @@ def test_quick_assoc():
     c, t = aes.quick_gcm_enc(b"A"*16, b"A"*16, b"Hello", assoc=b"blah")
     p = aes.quick_gcm_dec(b"A"*16, b"A"*16, c, t, assoc=b"blah")
     assert p == b"Hello"
-
-def test_ecb():
-        key = b"\x02" * 16
-        data = b"\x01" * 16
-
-        assert len(data) == 16
-        aes = Cipher("AES-128-ECB")
-        enc = aes.enc(key, None)
-        c = enc.update(data)
-        c += enc.finalize()
-
-        assert len(data) == 16
-        aes = Cipher("AES-128-ECB")
-        enc = aes.dec(key, None)
-        c1 = enc.update(c)
-        c1 += enc.finalize()
-        assert c1 == data
 
 # pylint: enable=unused-variable,redefined-outer-name
